Minor refactoring * Updated to use environment variable for InstalledSmartAppId * Including smartthings_rsa on slash command handling * smartapp to smartApp * (await).catch to try/catch * Fixed ;
