+++ src/lib/slack/commands/SlashCommand.js
@@ -26,26 +26,27 @@ const SlashCommand = {
     },
 
     sendResponse: async function(block) {
-        await request({
-	        method: 'POST',
-	        uri: this.responseUrl,
-	        json: true,
-	        body: block
-        }).catch((err) => {
+        try {
+            await request({
+                method: 'POST',
+                uri: this.responseUrl,
+                json: true,
+                body: block
+            })
+        } catch {
             console.error('ERROR ** Error posting to Slack response URL for slash command', err);
-        });
-    }
-}
-
-function requireParams(body, params) {
-    var validated = true;
-    for (var param of params) {
-        if (!body[param]) {
-            console.error(`ERROR ** Unable to init SlashCommand, no ${param} param found in body:`, body);
-            validated = false;
         }
     }
-    return validated;
+};
+
+function requireParams(body, params) {
+    for (const param of params) {
+        if (!body[param]) {
+            console.error(`ERROR ** Unable to init SlashCommand, no ${param} param found in body:`, body);
+            return false;
+        }
+    }
+    return true;
 }
 
 module.exports = SlashCommand;
