--- src/lib/smartthings/smartappService.js
@@ -1,148 +1,0 @@
-'use strict';
-
-const request = require('request-promise-native'),
-      smartapp = require('../smartApp');
-
-const smartappService = {
-    getSmartAppContext: async function(installedSmartAppId) {
-        var contextError;
-        const smartAppContext = await smartapp.withContext(installedSmartAppId)
-        .catch((err) => {
-            console.error(`Error creating SmartApp context for InstalledSmartApp ${installedSmartAppId}`, err);
-            contextError = err;
-        });
-        return [ contextError, smartAppContext ];
-    },
-
-    getDeviceList: async function(smartAppContext) {
-        var deviceListError;
-        const deviceList = await smartAppContext.api.devices.listAll()
-        .catch((err) => {
-            console.error(`ERROR ** Error listing devices for InstalledSmartApp ${installedSmartAppId}`, err);
-            deviceListError = err;
-        });
-        return [ deviceListError, deviceList ];
-    },
-
-    sendCommandSuccess: async function(responseUrl, command) {
-        await request({
-            method: 'POST',
-            uri: responseUrl,
-            json: true,
-            body: { "text": `Device command sent :thumbsup:` }
-        }).catch((err) => {
-            console.error(`ERROR ** Error posting to Slack response URL in ${command} command after a successful sendCommand`, err);
-        });
-    },
-
-    sendCommandError: async function(responseUrl, deviceId, command) {
-        await request({
-            method: 'POST',
-            uri: responseUrl,
-            json: true,
-            body: { "text": `Uh oh, there was an error trying to send the ${command} command to your device with ID ${deviceId}. Please try again.` }
-        }).catch((err) => {
-            console.error(`ERROR ** Error posting to Slack response URL for ${command} command after a failed sendCommand`, err);
-        });
-    },
-
-    configNotFoundErrorResponse: async function(responseUrl, deviceId, command) {
-        await request({
-            method: 'POST',
-            uri: responseUrl,
-            json: true,
-            body: { "text": `Oops, I wasn't able to find a device with the ID you specified for that command. Double check that the device with ID ${deviceId} was not deleted and then try again.` }
-        }).catch((err) => {
-            console.error(`ERROR ** Error posting to Slack response URL for ${command} command`, err);
-        });
-    }
-}
-
-const executableCommands = {
-    deviceOn: async function(deviceId, installedSmartAppId, responseUrl) {
-        const [ contextError, smartappContext ] = await this.getSmartAppContext(installedSmartAppId);
-        if (contextError) { return; };
-
-        const deviceConfig = smartappContext.config.lights.find(device => device.deviceConfig.deviceId == deviceId);
-        if (!deviceConfig) {
-            await this.configNotFoundErrorResponse(responseUrl, deviceId, 'on');
-            return;
-        }
-
-        var sendCommandError;
-        await smartappContext.api.devices.sendCommand(deviceConfig, 'switch', 'on')
-        .catch((err) => {
-            console.error(`Error sending on command to device ${deviceId}`, err);
-            sendCommandError = err;
-        });
-        if (sendCommandError) {
-            await this.sendCommandError(responseUrl, deviceId, 'on');
-            return;
-        }
-
-        await this.sendCommandSuccess(responseUrl, 'on');
-    },
-
-    deviceOff: async function(deviceId, installedSmartAppId, responseUrl) {
-        const [ contextError, smartappContext ] = await this.getSmartAppContext(installedSmartAppId);
-        if (contextError) { return; };
-
-        const deviceConfig = smartappContext.config.lights.find(device => device.deviceConfig.deviceId == deviceId);
-        if (!deviceConfig) {
-            await this.configNotFoundErrorResponse(responseUrl, deviceId, 'off');
-            return;
-        }
-
-        var sendCommandError;
-        await smartappContext.api.devices.sendCommand(deviceConfig, 'switch', 'off')
-        .catch((err) => {
-            console.error(`Error sending off command to device ${deviceId}`, err);
-            sendCommandError = err;
-        });
-        if (sendCommandError) {
-            await this.sendCommandError(responseUrl, deviceId, 'off');
-            return;
-        }
-        
-        await this.sendCommandSuccess(responseUrl, 'off');
-    },
-
-    deviceSetColor: async function(deviceId, color, installedSmartAppId, responseUrl) {
-        const [ contextError, smartappContext ] = await this.getSmartAppContext(installedSmartAppId);
-        if (contextError) { return; };
-
-        const deviceConfig = smartappContext.config.lights.find(device => device.deviceConfig.deviceId == deviceId);
-        if (!deviceConfig) {
-            await this.configNotFoundErrorResponse(responseUrl, deviceId, 'setColor');
-            return;
-        }
-
-        var sendCommandError;
-        await smartappContext.api.devices.sendCommand(
-            deviceConfig,
-            'colorControl',
-            'setColor',
-            [
-                {
-                    'hue': color,
-                    'saturation': 100,
-                    'level': 100,
-                    'switch': 'on'
-                }
-            ]
-        )
-        .catch((err) => {
-            console.error(`Error sending setColor command to device ${deviceId}`, err);
-            sendCommandError = err;
-        });
-        if (sendCommandError) {
-            await this.sendCommandError(responseUrl, deviceId, 'setColor');
-            return;
-        }
-        
-        await this.sendCommandSuccess(responseUrl, 'setColor');
-    }
-}
-
-Object.setPrototypeOf(smartappService, executableCommands);
-module.exports = smartappService;
